% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLumi.R
\name{shinyLumi}
\alias{shinyLumi}
\title{Launch shinyMethyl interactive session for shinyMethyl.}
\usage{
shinyLumi(Methylumi)
}
\arguments{
\item{Methylumi}{a MethylumiSet object contains methylation array information for plot generation}
}
\value{
No value returned. Instead the shinyMethyl interactive session is launched.
}
\description{
A functio that Convert MethylLumiSet (methylumi) to RGChannelSet (minfi), then generate plot use shinyMethyl
}
